+++ src/main/java/net/optifine/shaders/ClippingHelperShadow.java
@@ -32,12 +32,12 @@ public class ClippingHelperShadow extends ClippingHelper
 
     private double dot4(float[] plane, double x, double y, double z)
     {
-        return plane[0] * x + plane[1] * y + plane[2] * z + plane[3];
+        return (double)plane[0] * x + (double)plane[1] * y + (double)plane[2] * z + (double)plane[3];
     }
 
     private double dot3(float[] vecA, float[] vecB)
     {
-        return (double)vecA[0] * vecB[0] + (double)vecA[1] * vecB[1] + (double)vecA[2] * vecB[2];
+        return (double)vecA[0] * (double)vecB[0] + (double)vecA[1] * (double)vecB[1] + (double)vecA[2] * (double)vecB[2];
     }
 
     public static ClippingHelper getInstance()
@@ -71,7 +71,7 @@ public class ClippingHelperShadow extends ClippingHelper
 
     private void assignPlane(float[] plane, float a, float b, float c, float d)
     {
-        float f = (float)Math.sqrt(a * a + b * b + c * c);
+        float f = (float)Math.sqrt((double)(a * a + b * b + c * c));
         plane[0] = a / f;
         plane[1] = b / f;
         plane[2] = c / f;
@@ -100,7 +100,7 @@ public class ClippingHelperShadow extends ClippingHelper
 
     private float length(float x, float y, float z)
     {
-        return (float)Math.sqrt(x * x + y * y + z * z);
+        return (float)Math.sqrt((double)(x * x + y * y + z * z));
     }
 
     private float distance(float x1, float y1, float z1, float x2, float y2, float z2)
