+++ src/main/java/net/optifine/util/CompactArrayList.java
@@ -96,7 +96,7 @@ public class CompactArrayList
         }
         else if (this.list.size() > this.initialCapacity)
         {
-            float f = this.countValid * 1.0F / this.list.size();
+            float f = (float)this.countValid * 1.0F / (float)this.list.size();
 
             if (f <= this.loadFactor)
             {
